(refs #216) Change not to use field-types Revert "Merge pull request #376 from asha-nepal/reusable-form-definition/216"
This reverts commit 7c1ffb7f84588f58067c6d22c9efd89983a177f7, reversing
changes made to fbdd6e3c1e6d7febaa43c627f5340cf3aaa8fc95.
