+++ src/connects/PatientView.react.js
@@ -19,7 +19,6 @@ import {
   getRecordFormStyles,
   getRecordFormStyleId,
   getRecordFormStyle,
-  getRecordFieldTypes,
   getPatientFormStyle,
   getActiveRecords,
   getSelectedActiveRecordIndex,
@@ -45,7 +44,6 @@ const mapStateToProps = (state) => {
     recordFormStyles: getRecordFormStyles(state),
     recordFormStyleId: getRecordFormStyleId(state),
     recordFormStyle: getRecordFormStyle(state),
-    recordFieldTypes: getRecordFieldTypes(state),
     duplicatedPatientsExist: {
       name: getNameDuplicatedPatients(state).length > 0,
       number: getNumberDuplicatedPatients(state).length > 0,
