+++ src/store/reducers/dform/index.js
@@ -3,9 +3,7 @@ import {
 } from 'redux';
 
 import styles from './styles';
-import fieldTypes from './field-types';
 
 export default combineReducers({
   styles,
-  fieldTypes,
 });
